requirements.txt
python-telegram-bot
imageai~=2.1.6
